Renamed files to prepare for other tests that will go in here.
