+++ test/functional/forall/test-forall-segment-cuda.cpp
