+++ test/functional/forall/test-forall-segment-hip.cpp
